Revamp how the audio sources keep track of their AVCodecContexts to make it easier to assign them custom destructors
Take advantage of this to rewrite most of how the Haali audio source
reads file properties, moving all of the repeated code into a single
common function and storing all of the data in the AVCodecContext rather
than in several data structures.

Happens to fix Vorbis in containers opened by Haali's splitters
(vorbis_init is dumb and uses some properties from the codec context
before setting them, so they need to be set by the calling code).

Originally committed to SVN as r383.