Use library version from ffms.h in the .pc file
Unfortunately since autoconf is dumb and does a LOT of static
evaluation, I can't just put the version.sh script in AC_INIT. I did
put a call to esyscmd, but that only does static evaluation. A hack to
put the invocation in AC_INIT was attempted and worked but it broke
out-of-tree builds; a hackier hack to fix out-of-tree builds caused m4
to enter an infinite memory leaking loop of doom so that had to be
dropped.

A separate FFMS_VERSION variable was added to store the actual output of
version.sh at runtime, and is used for substitution in ffms2.pc.

The version.sh script also has some fallbacks in the case it can't use
awk or perl to extract version information; that should be incredibly
rare so I'm not sure whether that's dead code or whether it'll be
useful for anything.

Originally committed to SVN as r540.