Add support for WebM's reference-only frames
VP8 has frames which are used only as references for other frames and
are not part of the decoded output, which breaks the one packet in = one
frame out assumption. Add support for unframes to the index, and map
presentation frame numbers to decoded frame numbers.

Closes issue #88.

Originally committed to SVN as r677.