Fix a memleak, attempt 2.
av_bitstream_filter_filter() can reallocate the input buffer and give
you a different pointer back, and it's apparently your responsibility to
free it. This was previously not done.

The code is basically still incorrect and will not work for chains of
more than one bitstream filter, but I don't think that's ever been used
in ffms.

Originally committed to SVN as r741.