Symlink libffms2.so into $libdir/vapoursynth when installing.
With the new plugin autoloading feature in VapourSynth, this
relieves the user of having to manually copy or symlink FFMS2
into the autoloading directory themselves.

The symlinking only occurs on non-Windows OSes, since on Windows
there isn't any significance to the placement of ffms2.dll anyway
(and because true *nix-style symlinks on Windows are a dark art).