configure: Change default optlevel to -O2
It was set on 0 for the C-plugin branch only because the Matroska
demuxer tripped a nasty bug in GCC that would result in crashing
under certain circumstances.  Since the Matroska demuxer has been
removed, we no longer need to keep GCC optimizations disabled.

The master branch uses -O3, but this causes the C-plugin parts
to spark Access Violations in AviSynth when used through VirtualDub
and ffmpeg (although not mpv, strangely enough).