indexing: Don't bail on timestamp/duration-less packets if it's an alt-ref codec
VPx alt-refs are output as packets which lack timestmps or durations, since
they are invisible. Currently, the timestamp mangling code in libavformat
will sometimes add a bogus timestamp and duration, if the webm in question
has the block duration set (which is a hack, and not really a frame duration
at all). In the future, libav* will only output packets without timestamps
or duration, so we need to handle it here, regardless. This does not handle
NVOPs. We set the duration based on the last PTS, for these packets, because
FFMS2 currently sorts packets by PTS, which will break decoding, otherwise.

(ノ°Д°）ノ︵

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>