API: Add FFMS_Deinit
Because libavformat is great, and TLS libraries are great, and network
stacks are great, libavformat has some global state init functions that
do not need to be inited, and must only be run once, and also some other
global state init functions that must be deinited upon program exit,
and reinited if used again.

To this end, I've implemented this with a lock, so it can be re-inited
after deinit safely, if the user wants to do that for some reason.

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>