videosource: Initialize HDR metadata presence flags using actual data
The MasteringDisplaySideData and ContentLightSideData presence is not
guaranteed for every frame, as it is assumed that the validity of this
metadata lasts for the whole duration of the stream.

However the flags that check for their presence (HasMasteringDisplayPrimaries,
HasMasteringDisplayLuminance, and HasContentLightLevel) always use the
side data flags directly, even when they are not present. This leads to
an inconsistent LocalFrame state, in which HDR metadata is initialized,
but flags are reset, and callers can't use those fields.

Instead of relying on the side data flags, check for the presence of HDR
metadata directly, and initialize LocalFrame flags correctly.