ParseVP9: Don't mark frames as invisible in newer versions of FFmpeg
They changed how superframes are retured, so in older versions, they are split,
and you can get a packet that contains only an invisible frame, but in newer
versions, they are not split, and marking a whole superframe as invisible is
incorrect.

The version was not bumped for this breaking changed, of course,, so we choose
the next, unrelated version bump to check for.

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>