Add Unit Testing
This adds support within the build system for googletest, as well
as an inital test. googletest is added as submodule. Tested to work
both in-tree and out-of-tree.

Run with:

    $ .../configure
    $ make test-sync # if needed
    $ make test

Commands may also have the following make variables set via command line:

    * SAMPLES_URL
    * SAMPLES_DIR

Example:
    $ make test-sync SAMPLES_DIR=/home/user/sampels SAMPLES_URL=https://localsamples.com/samples

Syncing currently relies on wget.

This does not currently include MSVC project integration.

A simple tool is included for generating the test data. It can be run
as, for example:

    $ cd $SAMPLES_DIR
    $ .../gendata sample.mp4 samplemp4data 0 > .../data/sample.mp4.cpp

This tool is not comprehensive, and is just a starting point. Sample data
may need to be manually generated for some problem samples.

Initial-patch-by: Thierry Foucu <tfoucu@gmail.com>
Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>