Expose flip operations from the display matrix
It should be noted that previously the rotation value was used as is
and it was modified to be always positive by adding 360. However since
flip operations are exposed now, it is necessary to apply some more
logic and internally convert the rotation angle.

For non-flipped videos this should be a transparent change.

Fix #317.