Hard fail on IO error during indexing and decode
If we encounter an IO error while reading packets, we don't want
the current behavior of just ending the indexing or decoding here
and treating the file as truncated - we want to hard fail so it is
apparent there has been an IO issue.

This is helpful in e.g. networked situations.

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>