Fix FindClosestVideoKeyFrame
The previous logic would return the index of the closest keyframe
in decoding order, but the caller assumes it's the number of a frame
in output order. So the seeks would not actually be to presentation
timestamps of keyframes.

Add a PosInDecodingOrder field to frames to make calculations easier
and also document OriginalPos a bit more.

Also note that FindClosestVideoKeyframe(0) will now not always return 0,
e.g. when the first frame in output order isn't the first frame in
decoding order (in which case it returns the index in output order
of the first frame in decoding order).

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>