Don't default to UseDTS for mpegts
Presentation timestamps seem to work fine for most files in
these containers (and replacing PTS with DTS can break predictions)
so wait until we encounter a packet with no PTS (which does
happen sometimes) before reverting to UseDTS.

On the other hand, seeking by PTS is unreliable in mpegts (since
it's just lavf's generic internal seek function), so default to
SeekByPos there.

I don't have an mpegtsraw test file on hand but as far as I can see that
should get the same treatment. For nuv I don't really know, so I'll
leave it as it is.