Consider hidden frames in FrameFromPTS/Pos when seeking
01cae8948d809610c2f5864d40f18256aa09a8c6 added the AllowHidden parameter
to FrameFromPTS/Pos to be able to recognize the current packet in
DecodePacket, but kept it set to false in the other invocations to
change as little existing behavior as possible.

However, it looks like it's actually correct to also allow hidden frames
in the other invocations in GetFrame, since those call FrameFromPTS to
find out what packet the demuxer ended up at after a seek.

FFMS/ffms2#456 shows a file where this makes a difference, since its
first few frames are marked as hidden.

Setting AllowHidden to true everywhere in the decoding loop (just not in
GetFrameByTime) also causes no regressions on the Doom9 seeking sample
collection
(https://forum.doom9.org/showthread.php?p=1992736#post1992736).