Return a full AVPacket in DecodeNextFrame
Instead of writing the pts and pos to output args.
This does slightly change behavior: Previously, DecodeNextFrame would
return
- the pos of the first packet read, and
- the first *nonnegative* pts (or dts if Frames.UseDTS is true) of a
  packet read.
Hence, AStartTime and Pos could potentially belong to different packets
if the first packet read had a negative PTS, like in an mp4 whose start
is cut off by an edit list.
In such cases, returning the first nonnegative PTS is the necessary with
the current logic in GetFrame (which excludes hidden frames in the
FrameFromPTS call), so keep this logic for now. This will change in the
next commit, but this way this intermediate commit won't break any test
files.