Rework packet finding functions
After seeking, GetFrame() tries to find out where the demuxer ended up
after the seek by trying to recognize the first packet that comes out of
the demuxer afterwards by its PTS (or DTS if Frames.UseDTS is true), or
by its Pos if that fails.
Similarly, DecodePacket() needs to recognize the packet to see if it's a
second field of some interlaced frame.
When the same PTS can appear multiple times in the same file (e.g. in
files with edit lists), this can cause issues.

This commit replaces the FrameFromPTS / FrameFromPos functions and the
fallback logic in GetFrame() by a single FindPacket function that takes
an entire AVPacket.
FindPacket() tries to find the unique packet in the index whose
TS/Pos/Flags match the given packet. If no such packet exists,
it tries a fallback sequence of comparing fewer fields.

The current fallback sequence is mostly chosen based on intuition - for
well-behaved files simply checking all fields should work fine. If
examples come up where this fallback sequence becomes relevant, it can
be adjusted.